+++ CHANGELOG.md
@@ -2,9 +2,8 @@
 
 ## 0.22.4dev
 
-* [Feature] Support for selecting a custom executor in NotebookExecutor tasks
 * [Fix] Fix error in `ScriptRunner` that didn't allow to import modules in script's directory (#1072)
-
+* [Feature] Add `executor` option to NotebookRunner to choose ploomber-engine or papermill
 
 ## 0.22.3 (2023-04-13)